feature(/style/theme): 컬러 추가
- 간단한 main, sub 컬러만 추가
