+++ webrtc/src/public/js/app.js
@@ -1,4 +1,3 @@
-
 // const socket = new WebSocket(`ws://${window.location.host}`);
 // socket.addEventListener("open", ()=> {
 //     console.log("Connec");
